pyproject.toml
[project]
name = "seog"
version = "0.1.0"
description = "Simple Ensi OpenAPI Generator"
authors = [
    {name = "work.cortez13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "googletrans (==4.0.0-rc1)"
]

[project.scripts]
seog = "seog.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
